refact: parking the ChunkDataProvider in the provider module, will be refactored later
