+++ pathetic-provider/paper/pom.xml
@@ -33,5 +33,12 @@
             <version>3.3.1</version>
             <scope>provided</scope>
         </dependency>
+
+        <dependency>
+            <groupId>de.metaphoriker</groupId>
+            <artifactId>pathetic-provider</artifactId>
+            <version>3.3.1</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 </project>
