+++ pathetic-provider/spigot/pom.xml
@@ -31,6 +31,13 @@
 
         <dependency>
             <groupId>de.metaphoriker</groupId>
+            <artifactId>pathetic-provider</artifactId>
+            <version>3.3.1</version>
+            <scope>compile</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>de.metaphoriker</groupId>
             <artifactId>v1_21_R3</artifactId>
             <version>3.3.1</version>
             <scope>compile</scope>
