+++ pathetic-provider/spigot/v1_17/src/main/java/de/metaphoriker/pathetic/provider/v1_17/v1_17ChunkDataProviderImpl.java
@@ -1,5 +1,6 @@
 package de.metaphoriker.pathetic.provider.v1_17;
 
+import de.metaphoriker.pathetic.provider.ChunkDataProvider;
 import net.minecraft.server.level.WorldServer;
 import net.minecraft.world.level.block.state.IBlockData;
 import net.minecraft.world.level.chunk.ChunkStatus;
@@ -11,7 +12,6 @@ import org.bukkit.craftbukkit.v1_17_R1.CraftChunk;
 import org.bukkit.craftbukkit.v1_17_R1.CraftWorld;
 import org.bukkit.craftbukkit.v1_17_R1.block.CraftBlockStates;
 import org.bukkit.craftbukkit.v1_17_R1.block.data.CraftBlockData;
-import de.metaphoriker.pathetic.api.snapshot.ChunkDataProvider;
 
 public class v1_17ChunkDataProviderImpl implements ChunkDataProvider {
