+++ pathetic-provider/spigot/v1_20_R3/pom.xml
@@ -26,7 +26,12 @@
             <version>1.20.4-R0.1-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
-
+        <dependency>
+            <groupId>de.metaphoriker</groupId>
+            <artifactId>pathetic-provider</artifactId>
+            <version>3.3.1</version>
+            <scope>compile</scope>
+        </dependency>
         <dependency>
             <groupId>de.metaphoriker</groupId>
             <artifactId>pathetic-api</artifactId>
