+++ pathetic-provider/spigot/v1_20_R3/src/main/java/de/metaphoriker/pathetic/provider/v1_20_R3/v1_20_R3ChunkDataProviderImpl.java
@@ -1,6 +1,8 @@
 package de.metaphoriker.pathetic.provider.v1_20_R3;
 
 import java.lang.reflect.Field;
+
+import de.metaphoriker.pathetic.provider.ChunkDataProvider;
 import net.minecraft.server.level.WorldServer;
 import net.minecraft.world.level.block.state.IBlockData;
 import net.minecraft.world.level.chunk.ChunkStatus;
@@ -10,7 +12,6 @@ import org.bukkit.World;
 import org.bukkit.block.BlockState;
 import org.bukkit.craftbukkit.v1_20_R3.CraftChunk;
 import org.bukkit.craftbukkit.v1_20_R3.CraftWorld;
-import de.metaphoriker.pathetic.api.snapshot.ChunkDataProvider;
 
 public final class v1_20_R3ChunkDataProviderImpl implements ChunkDataProvider {
