+++ pathetic-provider/spigot/v1_20_R4/src/main/java/de/metaphoriker/pathetic/provider/v1_20_R4/v1_20_R4ChunkDataProviderImpl.java
@@ -1,5 +1,6 @@
 package de.metaphoriker.pathetic.provider.v1_20_R4;
 
+import de.metaphoriker.pathetic.provider.ChunkDataProvider;
 import net.minecraft.server.level.WorldServer;
 import net.minecraft.world.level.block.state.IBlockData;
 import net.minecraft.world.level.chunk.DataPaletteBlock;
@@ -9,7 +10,6 @@ import org.bukkit.World;
 import org.bukkit.block.BlockState;
 import org.bukkit.craftbukkit.v1_20_R4.CraftChunk;
 import org.bukkit.craftbukkit.v1_20_R4.CraftWorld;
-import de.metaphoriker.pathetic.api.snapshot.ChunkDataProvider;
 
 import java.lang.reflect.Field;
