+++ pathetic-provider/spigot/v1_21_R3/src/main/java/de/metaphoriker/pathetic/provider/v1_21_R3ChunkDataProviderImpl.java
@@ -1,6 +1,5 @@
 package de.metaphoriker.pathetic.provider;
 
-import de.metaphoriker.pathetic.api.snapshot.ChunkDataProvider;
 import net.minecraft.server.level.WorldServer;
 import net.minecraft.world.level.block.state.IBlockData;
 import net.minecraft.world.level.chunk.DataPaletteBlock;
